Organize checklist app structure
