+++ src/checklist/templates/index.html
@@ -52,7 +52,7 @@
             <div class="collapse navbar-collapse" id="top-navbar-1">
                 <ul class="nav navbar-nav navbar-right">                     
                     <li><a href="{% url 'login' %}">Login</a></li>
-                    <li><a href="{% url 'checklistForm' %}">Iniciar Checklist</a></li>
+                    <li><a href="{% url 'checklist:checklistForm' %}">Iniciar Checklist</a></li>
                     <li><a href="{% url 'view_pdf_cae' %}">Cartilha</a></li>
                 </ul>
             </div>
