+++ src/checklist/templates/success.html
@@ -1,0 +1,110 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+
+  <meta charset="utf-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <title>FISCAE - plataforma de apoio ao CAE </title>
+
+  <!-- CSS -->
+  <link rel="stylesheet" href="http://fonts.googleapis.com/css?family=Open+Sans:700,300,400">
+  <link rel="stylesheet" href="static/assets/bootstrap/css/bootstrap.min.css">
+  <link rel="stylesheet" href="static/assets/font-awesome/css/font-awesome.min.css">
+  <link rel="stylesheet" href="static/assets/elegant-font/code/style.css">
+  <link rel="stylesheet" href="static/assets/css/animate.css">
+  <link rel="stylesheet" href="static/assets/css/magnific-popup.css">
+  <link rel="stylesheet" href="static/assets/css/form-elements.css">
+  <link rel="stylesheet" href="static/assets/css/style.css">
+  <link rel="stylesheet" href="static/assets/css/media-queries.css">
+
+
+
+  <link rel="shortcut icon" sizes="20x20" href="static/assets/ico/f.png">
+  <link rel="apple-touch-icon-precomposed" sizes="144x144" href="static/assets/ico/apple-touch-icon-144-precomposed.png">
+  <link rel="apple-touch-icon-precomposed" sizes="114x114" href="static/assets/ico/apple-touch-icon-114-precomposed.png">
+  <link rel="apple-touch-icon-precomposed" sizes="72x72" href="static/assets/ico/apple-touch-icon-72-precomposed.png">
+  <link rel="apple-touch-icon-precomposed" href="static/assets/ico/apple-touch-icon-57-precomposed.png">
+
+</head>
+
+<body>
+
+  <!-- Loader -->
+  <div class="loader">
+    <div class="loader-img"></div>
+  </div>
+
+  <!-- Top menu -->
+  <nav class="navbar navbar-fixed-top" role="navigation">
+    <div class="container">
+      <div class="navbar-header">
+        <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#top-navbar-1">
+            <span class="sr-only">Toggle navigation</span>
+            <span class="icon-bar"></span>
+            <span class="icon-bar"></span>
+            <span class="icon-bar"></span>
+          </button>
+        <a href="/"> <img src="static/assets/img/logo/l.png" alt="Fiscae" class="img-responsive"></a>
+
+      </div>
+      <!-- Collect the nav links, forms, and other content for toggling -->
+      <div class="collapse navbar-collapse" id="top-navbar-1">
+        <ul class="nav navbar-nav navbar-right">
+          <li><a href="#top-content">Login</a></li>
+          <li><a href="{% url 'checklist:checklistForm' %}">Iniciar Checklist</a></li>
+          <li><a href="#what-we-do">Cartilha</a></li>
+          <li><a href="#contact">Denúncia</a></li>
+        </ul>
+      </div>
+    </div>
+  </nav>
+  
+  <!-- form Title -->
+  <div class="page-title-text wow fadeInUp animated" style="visibility: visible; animation-name: fadeInUp;"></div>
+      <div class="container"></div>
+         <div class="form-group section-container checklists-container"></div>
+                <h3></h3>
+        </div>
+      </div>
+  </div>
+               
+                       
+       <!-- What we do -->
+    <div class="block-3-container section-container what-we-do-container">
+        <div class="container">
+            <div class="form-group-item" style "visibility: visible;"></div>
+              <ul class="form-group"></ul>
+
+    <div id="personalizado">
+      <br><br><h3>Checklist preenchido com Sucesso!</h3><br><br>
+      
+      <br><br><a href="formChecklist">Voltar aos formulários</a><br><br>
+             
+            </div>
+          </div>
+        </div>
+    </div>
+
+
+
+    <!-- Javascript -->
+    <script src="static/assets/js/jquery-1.11.1.min.js"></script>
+    <script src="static/assets/bootstrap/js/bootstrap.min.js"></script>
+    <script src="static/assets/js/jquery.backstretch.min.js"></script>
+    <script src="static/assets/js/wow.min.js"></script>
+    <script src="static/assets/js/retina-1.1.0.min.js"></script>
+    <script src="static/assets/js/jquery.magnific-popup.min.js"></script>
+    <script src="static/assets/js/waypoints.min.js"></script>
+    <script src="static/assets/js/jquery.countTo.js"></script>
+    <script src="static/assets/js/masonry.pkgd.min.js"></script>
+    <script src="static/assets/js/scripts.js"></script>
+
+    <!--[if lt IE 10]>
+            <script src="static/assets/js/placeholder.js"></script>
+        <![endif]-->
+
+</body>
+
+</html>
