+++ src/checklist/tests.py
@@ -1,8 +1,8 @@
 from django.test import TestCase, Client
 from django.shortcuts import reverse
-from .models import School
+
 from .models import Question
-from .forms import ChecklistForm
+
 from django.contrib.auth.models import User
 
 
@@ -21,7 +21,10 @@ class ChecklistTest(TestCase):
         client = Client()
         client.login(username="amanda", password="123")
         data = {'checklist_type': 'TA'}
-        response = client.post(reverse('checklistForm'), data, follow=True)
+        response = client.post(
+                        reverse('checklist:checklistForm'),
+                        data, follow=True
+                        )
         print(response.redirect_chain)
         self.assertEquals(response.status_code, 200)
         self.assertEquals(response.redirect_chain, [('/answerForm', 302)])
@@ -30,14 +33,18 @@ class ChecklistTest(TestCase):
         client = Client()
         client.login(username="amanda", password="123")
         data = {'checklist_type': 'TT'}
-        response = client.post(reverse('checklistForm'), data, follow=True)
+        response = client.post(
+                        reverse('checklist:checklistForm'),
+                        data,
+                        follow=True
+                        )
         self.assertEquals(response.status_code, 200)
         self.assertEquals(response.redirect_chain, [])
 
     def testRenderChecklistForm(self):
         client = Client()
         client.login(username="amanda", password="123")
-        response = client.get(reverse('checklistForm'))
+        response = client.get(reverse('checklist:checklistForm'))
         self.assertEquals(response.status_code, 200)
