+++ src/merenda/urls.py
@@ -4,9 +4,12 @@ from merenda import views
 
 
 urlpatterns = [
+    url(r'^', include('checklist.urls', namespace='checklist')),
+    url(r'^$', views.index, name='index'),
+
     url(r'^admin/', admin.site.urls),
     url(r'^user/', include('user.urls')),
-    url(r'^$', views.index, name='index'),
+
     url(
         r'^view_pdf_cae/CartilhaCAE.pdf$',
         views.view_pdf_cae,
@@ -16,11 +19,4 @@ urlpatterns = [
     url(r'^findSchool', views.findSchool, name='findSchool'),
     url(r'^formSelect', views.formSelect, name='formSelect'),
     url(r'^tecForm', views.tecForm, name='tecForm'),
-    url(r'^viewChecklist', views.viewChecklist, name='viewChecklist'),
-
-    url(r'^Success', views.Success, name='Success'),
-    url(r'^notLoggedIn', views.notLoggedIn, name='notLoggedIn'),
-
-    url(r'^formChecklist', views.checklistForm, name='checklistForm'),
-    url(r'^answerForm', views.answerForm, name='answerForm'),
 ]
